+++ revrt/routing/base.py
@@ -1,9 +1,7 @@
-"""reVRt routing from a point to many points"""
+"""reVRt base routing functionality"""
 
 import json
-import time
 import logging
-from pathlib import Path
 from warnings import warn
 from functools import cached_property
 from itertools import pairwise
@@ -11,23 +9,16 @@ from itertools import pairwise
 import rasterio
 import numpy as np
 import xarray as xr
-import pandas as pd
 import dask.array as da
-import geopandas as gpd
 from shapely.geometry import Point
-from shapely.geometry import MultiLineString
 from shapely.geometry.linestring import LineString
 
-from revrt import RouteFinder, simplify_using_slopes
+from revrt import simplify_using_slopes
 from revrt.models.cost_layers import BarrierLayer
-from revrt.utilities.handlers import IncrementalWriter
-from revrt.utilities.monitoring import log_runtime
-from revrt.exceptions import (
-    revrtKeyError,
-    revrtLeastCostPathNotFoundError,
-    revrtRustError,
-)
+from revrt.routing.utilities import compute_lens
+from revrt.exceptions import revrtKeyError
 from revrt.warn import revrtWarning, revrtDeprecationWarning
+
 
 logger = logging.getLogger(__name__)
 LCP_AGG_COST_LAYER_NAME = "lcp_agg_costs"
@@ -486,7 +477,7 @@ class CharacterizedLayer:
                 f"{self.name}_length_km": 0,
             }
 
-        lens, __ = _compute_lens(route, cell_size)
+        lens, __ = compute_lens(route, cell_size)
 
         layer_data = getattr(layer_values, "data", layer_values)
         if not isinstance(layer_data, da.Array):  # pragma: no cover
@@ -676,551 +667,9 @@ class RouteMetrics:
 
     def _compute_path_length(self):
         """Compute the total length and cell by cell length of LCP"""
-        self.__lens, self._total_path_length = _compute_lens(
+        self.__lens, self._total_path_length = compute_lens(
             self._route_row_col, self.cell_size
         )
-
-
-class IncrementalRouteWriter(IncrementalWriter):
-    """Stream results to disk by appending each new result to a file
-
-    A new file is created if one does not exist.
-    """
-
-    def __init__(self, out_fp, crs=None):
-        """
-
-        Parameters
-        ----------
-        out_fp : path-like
-            Path to output file.
-        crs : rasterio.crs.CRS or dict, optional
-            Coordinate reference system for geometries when saving to
-            GeoPackage. By default, ``None``.
-        """
-        super().__init__(out_fp)
-        self.crs = crs
-
-    def preprocess_chunk(self, result):
-        """Turn result into a dataframe chunk
-
-        Parameters
-        ----------
-        result : dict
-            Route result dictionary as built by
-            ``RouteMetrics.compute()``.
-
-        Returns
-        -------
-        pandas.DataFrame or geopandas.GeoDataFrame
-            A dataframe holding the route result.
-        """
-        if "geometry" in result:
-            return gpd.GeoDataFrame(
-                [result], geometry="geometry", crs=self.crs
-            )
-        return pd.DataFrame([result])
-
-
-class BatchRouteProcessor:
-    """Class to manage batches of route computations"""
-
-    def __init__(
-        self,
-        routing_scenario,
-        route_definitions,
-        route_attrs=None,
-        mem_limit_gb=4,
-    ):
-        """
-
-        Parameters
-        ----------
-        routing_scenario : RoutingScenario
-            Scenario describing the cost layers and routing options.
-        route_definitions : Iterable
-            Sequence of ``(start_points, end_points, option)`` tuples
-            defining which points to route between. Each of
-            ``start_points`` and ``end_points`` should be a list of
-            ``(row, col, option_name)`` tuples.
-        route_attrs : dict, optional
-            Mapping of tuples of the form (int, (int, int, option))
-            where the first integer represents the route ID and the
-            tuple of integers + str represents the starting index to
-            additional attributes to include in the output for that
-            route. By default, ``None``.
-        mem_limit_gb : int or float, default=4
-            Memory limit in gigabytes for routing computations.
-            By default, ``4``.
-        """
-        self.routing_scenario = routing_scenario
-        self._route_definitions = route_definitions
-        self._route_attrs = route_attrs or {}
-        self.mem_limit_gb = mem_limit_gb
-
-    @cached_property
-    def default_attrs(self):
-        """dict: Default attributes for all routes"""
-        keys = set().union(*[set(x) for x in self._route_attrs.values()])
-        return dict.fromkeys(keys)
-
-    @cached_property
-    def route_attrs(self):
-        """dict: Mapping of frozen route node pair sets to attributes"""
-        return {
-            k: {**self.default_attrs, **v}
-            for k, v in self._route_attrs.items()
-        }
-
-    @cached_property
-    def route_definitions(self):
-        """list: Validated route definitions for computation"""
-        return self._compile_valid_route_definitions()
-
-    @cached_property
-    def routing_layers(self):
-        """RoutingLayerManager: Built routing layers for the scenario"""
-        return RoutingLayerManager(self.routing_scenario).build()
-
-    def process(self, out_fp, save_paths=False, routing_layer_out_fp=None):
-        """Compute all routes and save to disk
-
-        Parameters
-        ----------
-        out_fp : path-like
-            Path to output file. If ``save_paths=True``, a GeoPackage
-            will be created (recommend to pass in a filepath ending in
-            ".gpkg"). Otherwise, a CSV file will be created (recommend
-            to pass in a filepath ending in ".csv").
-        save_paths : bool, default=False
-            Include shapely geometries in the output when ``True``.
-            By default, ``False``.
-        routing_layer_out_fp : path-like, optional
-            Optional output path for Rust routing-layer cache data.
-            By default, ``None``.
-        """
-        if not self.route_definitions:
-            return
-
-        with log_runtime(
-            f"Routing for {len(self.route_definitions)} route definitions"
-        ):
-            try:
-                self._compute_routes(
-                    out_fp, save_paths=save_paths, rl=routing_layer_out_fp
-                )
-            finally:
-                self._reset_routing_layers()
-
-    def _compute_routes(self, out_fp, save_paths, rl=None):
-        """Evaluate route definitions and build result records"""
-
-        out_fp = _validate_out_fp(out_fp, save_paths)
-        writer = IncrementalRouteWriter(
-            out_fp, crs=self.routing_layers.cost_crs
-        )
-        option_writer = None
-        if save_paths and self.routing_scenario.routing_options is not None:
-            option_writer = IncrementalRouteWriter(
-                _routing_options_output_fp(out_fp),
-                crs=self.routing_layers.cost_crs,
-            )
-
-        for indices, optimized_objective, attrs in self._route_results(rl):
-            metrics = RouteMetrics(
-                self.routing_layers,
-                indices,
-                optimized_objective,
-                add_geom=save_paths,
-                attrs=attrs,
-            )
-            route_result = metrics.compute()
-            writer.save(route_result)
-            if option_writer is not None:
-                for option_result in self._routing_option_results(
-                    indices, route_result
-                ):
-                    option_writer.save(option_result)
-
-    def _routing_option_results(self, indices, route_result):
-        """Yield aggregated geometries for each routing option used"""
-
-        segments_by_option = {}
-        current_option = None
-        current_segment = []
-        for start_p, end_p in pairwise(indices):
-            if start_p == end_p:
-                continue
-
-            start = tuple(start_p[:2])
-            end = tuple(end_p[:2])
-            start_point_option = start_p[-1]
-            if (
-                current_option == start_point_option
-                and current_segment[-1] == start
-            ):
-                current_segment.append(end)
-                continue
-
-            if current_segment:
-                segments_by_option.setdefault(current_option, []).append(
-                    current_segment
-                )
-            current_option = start_point_option
-            current_segment = [start, end]
-
-        if current_segment:
-            segments_by_option.setdefault(current_option, []).append(
-                current_segment
-            )
-
-        cell_size = abs(self.routing_layers.transform.a)
-        results = []
-        for option, segments in segments_by_option.items():
-            geoms = [self._component_geometry(segment) for segment in segments]
-            if not geoms:
-                continue
-
-            geometry = (
-                geoms[0]
-                if len(geoms) == 1
-                else MultiLineString([list(geom.coords) for geom in geoms])
-            )
-            length_km = sum(
-                _compute_lens(segment, cell_size)[1] for segment in segments
-            )
-            results.append(
-                {
-                    **{
-                        key: value
-                        for key, value in route_result.items()
-                        if key
-                        not in {
-                            "geometry",
-                            "cost",
-                            "optimized_objective",
-                            "length_km",
-                        }
-                    },
-                    "routing_option": option,
-                    "length_km": length_km,
-                    "geometry": geometry,
-                }
-            )
-
-        return results
-
-    def _component_geometry(self, route):
-        """Build geometry for one contiguous routing-option segment"""
-        rows, cols = np.array(route).T
-        x, y = rasterio.transform.xy(self.routing_layers.transform, rows, cols)
-        if len(route) == 1:
-            return Point(x, y)
-
-        return LineString(simplify_using_slopes(list(zip(x, y, strict=True))))
-
-    def _route_results(self, routing_layer_out_fp=None):
-        """Generator yielding route results from Rust computations"""
-        if not self.route_definitions:
-            return
-
-        logger.debug(
-            "Setting memory limit to %.2f GB for Rust computations",
-            self.mem_limit_gb,
-        )
-        route_results = RouteFinder(
-            zarr_fp=self.routing_scenario.cost_fpath,
-            cost_function=self.routing_scenario.cost_function_json,
-            route_definitions=[
-                (rid, sp, ep)
-                for rid, (sp, ep) in self.route_definitions.items()
-            ],
-            mem_limit_bytes=int(self.mem_limit_gb * 1_000_000_000),
-            algorithm=self.routing_scenario.algorithm,
-            log_level=logging.getLogger("revrt").level or None,
-            routing_layer_out_fp=routing_layer_out_fp,
-        )
-        yield from self._skip_failed_routes(route_results)
-
-    def _compile_valid_route_definitions(self):
-        """Filter route definitions to those with valid route nodes"""
-        if not self._route_definitions:
-            return {}
-
-        sample_definition = self._route_definitions[0]
-        if len(sample_definition) == 2:  # noqa: PLR2004
-            self._route_definitions = _add_route_ids(self._route_definitions)
-
-        routes_to_compute = {}
-        for route_id, start_points, end_points in self._route_definitions:
-            filtered_start_points = self._validate_start_points(start_points)
-            if not filtered_start_points:
-                msg = (
-                    f"All start points are invalid for route with ID "
-                    f"{route_id}: {start_points}\nSkipping..."
-                )
-                warn(msg, revrtWarning)
-                continue
-
-            try:
-                filtered_end_points = self._validate_end_points(end_points)
-            except revrtLeastCostPathNotFoundError:
-                continue
-
-            if not filtered_end_points:
-                msg = (
-                    f"All end points are invalid for route with ID "
-                    f"{route_id}: {end_points}\nSkipping..."
-                )
-                warn(msg, revrtWarning)
-                continue
-
-            routes_to_compute[route_id] = (
-                filtered_start_points,
-                filtered_end_points,
-            )
-
-        return routes_to_compute
-
-    def _skip_failed_routes(self, routing_results):
-        """Yield only successfully computed routes from Rust results"""
-
-        results_iter = iter(routing_results)
-        num_complete = 0
-        ts = time.monotonic()
-        while True:
-            num_complete += 1
-            try:
-                route_id, solutions = next(results_iter)
-                yield from self._formatted_solutions(solutions, route_id)
-                time_elapsed = f"{(time.monotonic() - ts) / 60:.2f} minute(s)"
-                logger.info(
-                    "%d/%d (%.2f%%) route definitions processed in %s",
-                    num_complete,
-                    len(self.route_definitions),
-                    (num_complete / len(self.route_definitions)) * 100,
-                    time_elapsed,
-                )
-            except revrtRustError:  # pragma: no cover
-                logger.exception("Rust error when computing route")
-                continue
-            except StopIteration:
-                logger.info("Routing complete")
-                break
-
-    def _formatted_solutions(self, solutions, route_id):
-        """Format reVRt output solutions and log any failures"""
-        start_points, end_points = self.route_definitions[route_id]
-        if not solutions:
-            msg = (
-                f"Unable to find route from {start_points} to any of "
-                f"{end_points} (route ID: {route_id}). Please verify "
-                "that the start and end points are not separated by "
-                "hard barriers or invalid cost cells."
-            )
-            logger.error(msg)
-            return
-
-        logger.debug(
-            "Got result from Rust for route_id %d. Processing..."
-            "\n\t- Start points: %r\n\t- End points: %r",
-            route_id,
-            start_points,
-            end_points,
-        )
-        for solution in solutions:
-            if len(solution) == ROUTE_SOLUTION_LEN:
-                indices, optimized_objective, dbl = solution
-            else:  # pragma: no cover
-                msg = f"Unexpected route solution payload: {solution!r}"
-                raise revrtKeyError(msg)
-
-            attrs_key = (route_id, indices[0])
-            attrs = {
-                **self.route_attrs.get(attrs_key, self.default_attrs),
-                "dropped_barrier_layers": json.dumps(dbl),
-            }
-            yield indices, optimized_objective, attrs
-
-    def _validate_start_points(self, points):
-        """Validate start points by removing cells invalid cost"""
-        points = _get_valid_points(
-            points, self.routing_layers.full_shape, point_type="start"
-        )
-        if not points or not self.routing_scenario.invalid_costs_block_routing:
-            return points
-
-        routing_options = {point[-1] for point in points}
-        bad_point_inds = set()
-        for r_o in routing_options:
-            rows, cols = np.array(
-                [point[:2] for point in points if point[-1] == r_o]
-            ).T
-            costs = self.routing_layers.costs[r_o].isel(
-                y=xr.DataArray(rows, dims="points"),
-                x=xr.DataArray(cols, dims="points"),
-            )
-
-            cost_values = costs.compute()
-            bad_point_inds |= set(
-                np.where(np.isnan(cost_values) | (cost_values <= 0))[0]
-            )
-
-        if not bad_point_inds:
-            return points
-
-        invalid_points = {points[i] for i in bad_point_inds}
-        msg = (
-            f"One or more of the start points have an invalid cost "
-            f"(must be > 0): {invalid_points}\n"
-            "Dropping these from consideration..."
-        )
-        warn(msg, revrtWarning)
-
-        if not points:
-            all_invalid_points_msg = (
-                "None of the start points have a valid cost (must be > 0): "
-                f"{points}"
-            )
-            raise revrtLeastCostPathNotFoundError(all_invalid_points_msg)
-
-        return [p for p in points if p not in invalid_points]
-
-    def _validate_end_points(self, points):
-        """Filter out invalid endpoints; raise if all are invalid"""
-        points = _get_valid_points(
-            points, self.routing_layers.full_shape, point_type="end"
-        )
-        if not points or not self.routing_scenario.invalid_costs_block_routing:
-            return points
-
-        routing_options = {point[-1] for point in points}
-        bad_point_inds = set()
-        for r_o in routing_options:
-            rows, cols = np.array(
-                [point[:2] for point in points if point[-1] == r_o]
-            ).T
-            costs = self.routing_layers.costs[r_o].isel(
-                y=xr.DataArray(rows, dims="points"),
-                x=xr.DataArray(cols, dims="points"),
-            )
-
-            cost_values = costs.compute()
-            bad_point_inds |= set(
-                np.where(np.isnan(cost_values) | (cost_values <= 0))[0]
-            )
-        if not bad_point_inds:
-            return points
-
-        invalid_points = {points[i] for i in bad_point_inds}
-        msg = (
-            f"One or more of the end points have an invalid cost "
-            f"(must be > 0): {invalid_points}\n"
-            "Dropping these from consideration..."
-        )
-        warn(msg, revrtWarning)
-        points = [p for p in points if p not in invalid_points]
-
-        if not points:
-            all_invalid_points_msg = (
-                "None of the end points have a valid cost (must be > 0): "
-                f"{sorted(invalid_points)}"
-            )
-            raise revrtLeastCostPathNotFoundError(all_invalid_points_msg)
-
-        return points
-
-    def _reset_routing_layers(self):
-        """Close handler and remove built routing layers from memory"""
-        self.routing_layers.close()
-        del self.routing_layers
-
-
-def _validate_out_fp(out_fp, save_paths):
-    """Validate output filepath extension"""
-    out_fp = Path(out_fp)
-
-    if save_paths and out_fp.suffix.lower() != ".gpkg":
-        msg = (
-            "When saving paths, the output file should have a '.gpkg' "
-            f"extension to ensure proper format! Got input file: '{out_fp}'. "
-            "Adding '.gpkg' extension... "
-        )
-        warn(msg, revrtWarning)
-        out_fp = out_fp.with_suffix(".gpkg")
-    elif not save_paths and out_fp.suffix.lower() != ".csv":
-        msg = (
-            "When not saving paths, the output file should have a '.csv' "
-            f"extension to ensure proper format! Got input file: '{out_fp}'. "
-            "Adding '.csv' extension... "
-        )
-        warn(msg, revrtWarning)
-        out_fp = out_fp.with_suffix(".csv")
-
-    logger.debug("Validated output filepath: %s", out_fp)
-    return out_fp
-
-
-def _get_valid_points(points, arr_shape, point_type):
-    """Get only points that are within array bounds"""
-    valid_points = []
-    invalid_points = []
-    for point in points:
-        if _is_valid_point(point, arr_shape):
-            valid_points.append(point)
-        else:
-            invalid_points.append(point)
-
-    if invalid_points:
-        msg = (
-            f"One or more of the {point_type} points are out of bounds for an "
-            f"array of shape {arr_shape}: {invalid_points}\n"
-            "Dropping these from consideration..."
-        )
-        warn(msg, revrtWarning)
-
-    return valid_points
-
-
-def _is_valid_point(point, arr_shape):
-    """Check if point is within array bounds"""
-    row, col, *__ = point
-    return 0 <= row < arr_shape[0] and 0 <= col < arr_shape[1]
-
-
-def _add_route_ids(route_definitions):
-    """Add route IDs to route definitions missing them"""
-    logger.info(
-        "Route ID's missing from route definitions - adding definition "
-        "index as route ID..."
-    )
-    return [
-        (ind, start_points, end_points)
-        for ind, (start_points, end_points) in enumerate(route_definitions)
-    ]
-
-
-def _compute_lens(route, cell_size):
-    """Compute the total length and cell by cell length of LCP"""
-    # Use Pythagorean theorem to calculate length between cells (km)
-    # Use c**2 = a**2 + b**2 to determine length of individual paths
-    lens = np.sqrt(np.sum(np.diff(route, axis=0) ** 2, axis=1))
-    total_path_length = np.sum(lens) * cell_size / 1000
-
-    # Need to determine distance coming into and out of any cell.
-    # Assume paths start and end at the center of a cell. Therefore,
-    # distance traveled in the cell is half the distance entering it
-    # and half the distance exiting it. Duplicate all lengths,
-    # pad 0s on ends for start  and end cells, and divide all
-    # distance by half.
-    lens = np.repeat(lens, 2)
-    lens = np.insert(np.append(lens, 0), 0, 0)
-    lens /= 2
-
-    # Group entrance and exits distance together, and add them
-    lens = lens.reshape((int(lens.shape[0] / 2), 2))
-    lens = np.sum(lens, axis=1)
-    return lens, total_path_length
 
 
 def _transition_cost_lookup(transition_costs):
@@ -1238,11 +687,6 @@ def _transition_cost_lookup(transition_costs):
         pairwise_costs[(dst, src)] = rule["cost"]
 
     return default_cost, pairwise_costs
-
-
-def _routing_options_output_fp(out_fp):
-    """pathlib.Path: Companion output path for routing-option pieces"""
-    return out_fp.with_name(f"{out_fp.stem}_routing_options{out_fp.suffix}")
 
 
 def _cost_layers_for_rust(layers):
